Move FormPage to payment-element-test-pages
